requirements.txt
Django
south
mysql-python
pyyaml